refactor(chat): 도메인별 mapper 분리
- 기존: ChatMapper가 모든 변환 수행
- 현재: ChatMapper, FileMapper, PageMapper로 분리
- ChatMessage와 ChatRoom은 밀접하게 연관되어 있어 ChatMapper에서 일괄 처리
- service, test에 변경 사항 반영
