+++ communication-service/src/main/java/com/example/communicationservice/mapper/ChatMapper.java
@@ -1,16 +1,13 @@
 package com.example.communicationservice.mapper;
 
-import com.example.communicationservice.client.dto.output.FileDownloadUrlGenerateOutput;
 import com.example.communicationservice.client.dto.output.FileDownloadUrlListGenerateOutput;
-import com.example.communicationservice.client.dto.output.FileUploadUrlGenerateOutput;
-import com.example.communicationservice.controller.dto.FileInfo;
-import com.example.communicationservice.controller.dto.request.ChatFileSendRequest;
 import com.example.communicationservice.controller.dto.request.ChatMessageSendRequest;
-import com.example.communicationservice.controller.dto.response.*;
+import com.example.communicationservice.controller.dto.response.ChatMessageReadResponse;
+import com.example.communicationservice.controller.dto.response.ChatMessageSendResponse;
+import com.example.communicationservice.controller.dto.response.ChatRoomCreateResponse;
+import com.example.communicationservice.controller.dto.response.ChatRoomReadResponse;
 import com.example.communicationservice.entity.ChatMessage;
 import com.example.communicationservice.entity.ChatRoom;
-import com.example.communicationservice.entity.File;
-import org.springframework.data.domain.Page;
 
 // dto <-> entity 또는 dto <-> dto 변환 로직 전담
 public abstract class ChatMapper {
@@ -25,7 +22,7 @@ public abstract class ChatMapper {
             .senderCode(request.senderCode())
             .type(request.type())
             .text(request.text())
-            .file(request.file() != null ? toEntity(request.file()) : null)
+            .file(request.file() != null ? FileMapper.toEntity(request.file()) : null)
             .build();
     }
 
@@ -35,7 +32,7 @@ public abstract class ChatMapper {
             message.getSenderCode(),
             message.getType(),
             message.getText(),
-            from(message.getFile()),
+            FileMapper.from(message.getFile()),
             message.getSentAt()
         );
     }
@@ -50,7 +47,7 @@ public abstract class ChatMapper {
             message.getSenderCode(),
             message.getType(),
             message.getText(),
-            output != null ? from(output.urls().get(0)) : null,
+            output != null ? FileMapper.from(output.urls().get(0)) : null,
             message.getSentAt()
         );
     }
@@ -66,53 +63,6 @@ public abstract class ChatMapper {
             chatRoom.getId(),
             chatRoom.getName(),
             chatRoom.getUpdatedAt()
-        );
-    }
-
-    // ------------------ Paging ------------------
-
-    public static PageInfo toPageInfo(Page<?> page) {
-        return new PageInfo(
-            page.getNumber(),
-            page.getSize(),
-            page.getTotalElements(),
-            page.getTotalPages(),
-            page.hasNext()
-        );
-    }
-
-    // ------------------ File ------------------
-
-    public static FileInfo from(File file) {
-        if (file == null) {
-            return null;
-        }
-
-        return new FileInfo(file.getKey());
-    }
-
-    public static File toEntity(ChatFileSendRequest request) {
-        if (request == null) {
-            return null;
-        }
-
-        return File.builder()
-            .key(request.key())
-            .build();
-    }
-
-    public static ChatFileUploadUrlGenerateResponse from(FileUploadUrlGenerateOutput output) {
-        return new ChatFileUploadUrlGenerateResponse(
-            output.key(),
-            output.queryString()
-        );
-    }
-
-    public static ChatFileSendResponse from(FileDownloadUrlGenerateOutput output) {
-        return new ChatFileSendResponse(
-            output.key(),
-            output.queryString(),
-            output.fileType()
         );
     }
